requirements.txt
pdfplumber==0.9.0
Pillow==10.0.0
pyzbar==0.1.9
